Use postgresql['dir'] directory when it exists

Closes #7233

See merge request https://gitlab.com/gitlab-org/omnibus-gitlab/-/merge_requests/8203

Merged-by: Ian Baum <ibaum@gitlab.com>
Approved-by: Douglas Barbosa Alexandre <dbalexandre@gmail.com>
Approved-by: Ian Baum <ibaum@gitlab.com>

+++ files/gitlab-ctl-commands-ee/lib/geo/pitr_file.rb
@@ -5,15 +5,20 @@ require_relative '../consul'
 
 module Geo
   class PitrFile
-    attr_accessor :filepath, :consul_key
+    attr_accessor :filepath, :consul_key, :fallback_filepath
 
     NAMESPACE = 'geo/pitr'
+    PITR_FILE_NAME = 'geo-pitr-file'
+    CONSUL_PITR_KEY = 'promote-db'
 
-    # @param [String] filepath a full path for a on disk PITR file
-    # @param [String] consul_key a consul key to be instead used when consul is available
-    def initialize(filepath, consul_key:)
-      @filepath = filepath
-      @consul_key = "#{NAMESPACE}/#{consul_key}".downcase
+    # @param [Object] ctl the omnibus-ctl object associated with the command
+    def initialize(ctl)
+      @filepath = "#{GitlabCtl::Util.get_public_node_attributes.dig('postgresql', 'dir')}/data/#{PITR_FILE_NAME}"
+      @consul_key = "#{NAMESPACE}/#{CONSUL_PITR_KEY}".downcase
+
+      # TODO: https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8970
+      # Remove the line below after required stop, to ensure that nobody will be using the fallback path anymore
+      @fallback_filepath = "#{ctl.data_path}/postgresql/data/#{PITR_FILE_NAME}"
     end
 
     def use_consul?
@@ -29,13 +34,18 @@ module Geo
       else
         File.write(filepath, lsn)
       end
+    rescue Errno::ENOENT, ConsulHandler::ConsulError
+      raise PitrFileError, "Unable to create PITR"
     end
 
     def delete
       if use_consul?
         ConsulHandler::Kv.delete(consul_key)
-      elsif File.exist?(filepath)
-        File.delete(filepath)
+      else
+        File.delete(filepath) if File.exist?(filepath)
+        # TODO: https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8970
+        # Remove the line below after required stop, to ensure that nobody will be using the fallback path anymore
+        File.delete(fallback_filepath) if File.exist?(fallback_filepath)
       end
     rescue Errno::ENOENT, ConsulHandler::ConsulError
       raise PitrFileError, "Unable to delete PITR"
@@ -44,8 +54,13 @@ module Geo
     def get
       if use_consul?
         ConsulHandler::Kv.get(consul_key)
-      else
+      elsif File.exist?(filepath)
         File.read(filepath)
+      else
+        # TODO: https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/8970
+        # Remove the line below as well as the check for file existence above after required stop,
+        # to ensure that nobody will be using the fallback path anymore
+        File.read(fallback_filepath)
       end
     rescue Errno::ENOENT, ConsulHandler::ConsulError
       raise PitrFileError, "Unable to fetch PITR"
